requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
class-registry==2.1.2
filters==1.3.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.2
regex==2024.9.11
six==1.16.0
sniffio==1.3.1
